+++ src/admin/resources/js/cmsadmin.js
@@ -1173,6 +1173,7 @@
 			typeDataCopy.keywords = itemCopy.keywords;
 			typeDataCopy.timestamp_create = itemCopy.timestamp_create;
 			typeDataCopy.image_id = itemCopy.image_id;
+			typeDataCopy.is_url_strict_parsing_disabled = itemCopy.is_url_strict_parsing_disabled;
 			$http.post(
 				'admin/api-cms-navitem/update-page-item?navItemId=' + navItemId + '&navItemType=' + itemCopy.nav_item_type,
 				$.param(typeDataCopy),