+++ CHANGELOG.md
@@ -11,6 +11,7 @@ In order to read more about upgrading and BC breaks have a look at the [UPGRADE 
 
 ### Added
 
++ [#51](https://github.com/luyadev/luya-module-cms/issues/51) Enable strict url parsing as long as the previous page is a module or strict parsing is disabled.
 + [#82](https://github.com/luyadev/luya-module-cms/issues/82) Option to define target (new window or not) for redirect pages.
 + [#159](https://github.com/luyadev/luya-module-cms/issues/159) Added new og:image and twitter:image meta informations. Its possible now to upload an image for a page.